Updates to build 186, fixes imports and code style issues.

+++ src/main/java/io/tabletoptools/hawthorne/listener/MessageListener.java
@@ -1,11 +1,11 @@
 /*
  * $Id: HttpFilter 3988 2017-06-21 13:47:09Z cfi $
  * Created on 05.02.18 14:23
- * 
+ *
  * Copyright (c) 2017 by bluesky IT-Solutions AG,
  * Kaspar-Pfeiffer-Strasse 4, 4142 Muenchenstein, Switzerland.
  * All rights reserved.
- * 
+ *
  * This software is the confidential and proprietary information
  * of bluesky IT-Solutions AG ("Confidential Information").  You
  * shall not disclose such Confidential Information and shall use
@@ -16,27 +16,20 @@ package io.tabletoptools.hawthorne.listener;
 
 import ch.hive.discord.bots.commands.CommandBase;
 import io.tabletoptools.discord.modulizer.Modulizer;
-import io.tabletoptools.hawthorne.HawthorneBot;
 import io.tabletoptools.hawthorne.model.ListMessageInstance;
 import io.tabletoptools.hawthorne.model.LookupItem;
 import io.tabletoptools.hawthorne.services.HomebrewItemService;
 import io.tabletoptools.hawthorne.util.SearchUtils;
-import me.xdrop.fuzzywuzzy.FuzzySearch;
-import me.xdrop.fuzzywuzzy.model.ExtractedResult;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.MessageBuilder;
 import net.dv8tion.jda.core.entities.Message;
 import net.dv8tion.jda.core.entities.MessageEmbed;
-import net.dv8tion.jda.core.entities.User;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.core.hooks.ListenerAdapter;
-import net.dv8tion.jda.core.requests.restaction.MessageAction;
 
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.List;
 
 public class MessageListener extends ListenerAdapter {
-
 
 
     @Override
@@ -50,7 +43,7 @@ public class MessageListener extends ListenerAdapter {
                     .setTitle("Audit Request")
                     .setAuthor(event.getAuthor().getName(), null, event.getAuthor().getEffectiveAvatarUrl())
                     .setDescription(new StringBuilder()
-                        .append(event.getAuthor().getName())
+                            .append(event.getAuthor().getName())
                             .append(" would like to get audited.")
                     )
                     .setFooter(event.getAuthor().getName() +
@@ -92,18 +85,17 @@ public class MessageListener extends ListenerAdapter {
                         .getInstances(event.getChannel().getIdLong(), event.getAuthor().getIdLong())
                         .forEach(instance -> instance.command(event, event.getMessage().getContentRaw()));
             }
-        } else if (event.getMessage().getContentRaw().matches("^[1-8]$")) {
-            if (ListMessageInstance.hasInstance(event.getChannel().getIdLong(), event.getAuthor().getIdLong())) {
-                ListMessageInstance
-                        .getInstances(event.getChannel().getIdLong(), event.getAuthor().getIdLong())
-                        .forEach(instance -> event.getChannel()
-                                .sendMessage(instance.choose(Integer.parseInt(event.getMessage().getContentRaw())))
-                                .queue());
-                event.getMessage().delete().queue();
-            }
+        } else if (event.getMessage().getContentRaw().matches("^[1-8]$")
+                && ListMessageInstance.hasInstance(event.getChannel().getIdLong(), event.getAuthor().getIdLong())) {
+            ListMessageInstance
+                    .getInstances(event.getChannel().getIdLong(), event.getAuthor().getIdLong())
+                    .forEach(instance -> event.getChannel()
+                            .sendMessage(instance.choose(Integer.parseInt(event.getMessage().getContentRaw())))
+                            .queue());
+            event.getMessage().delete().queue();
         }
 
-        if(event.getAuthor().getIdLong() == 135468266695950336L) {
+        if (event.getAuthor().getIdLong() == 135468266695950336L) {
 
             Modulizer.instance().process(event);
         }
